+++ src/components/CopyUiProvider/index.ts
@@ -1,4 +1,5 @@
-// CopyUiProvider from copy-ui @ 2025-10-27.
+// CopyUiProvider from copy-ui @ 2025-11-06.
 
 export { CopyUiProvider } from "./copy-ui-provider";
 export { useJss } from "./use-jss";
+export { useTheme } from "./use-theme";